authentication fix

+++ app/database.py
@@ -1,6 +1,8 @@
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import create_async_engine, async_sessionmaker
 
+from fastapi import HTTPException, status
+
 from app.models import UserOrm, Base, PostOrm
 from app.config import settings
 from app.schemas import UserCreate, PostCreate, UserUpdate, PostUpdate
@@ -30,8 +32,12 @@ async def get_all_users():
 
 async def get_user(user_id: int):
     async with session_factory() as session:
-        result = await session.execute(select(UserOrm).filter(UserOrm.id == user_id))
-        return result.scalar_one_or_none()
+        user = await session.get(UserOrm, user_id)
+        if user is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f"User not found")
+        return user
 
 
 async def get_user_by_username(username: str):
@@ -52,7 +58,7 @@ async def get_post(post_id: int):
         return result.scalar_one_or_none()
 
 
-async def insert_user(user: UserCreate):
+async def create_user(user: UserCreate):
     user_to_add = UserOrm(
         username=user.username,
         email=user.email,
@@ -66,6 +72,10 @@ async def insert_user(user: UserCreate):
 async def update_user(user_id: int, user_data: UserUpdate):
     async with session_factory() as session:
         user = await session.get(UserOrm, user_id)
+        if user is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f"User not found")
         user.username = user_data.username or user.username
         user.email = user_data.email or user.email
         user.password = user_data.password or user.password
@@ -74,7 +84,7 @@ async def update_user(user_id: int, user_data: UserUpdate):
         return user
 
 
-async def insert_post(post: PostCreate):
+async def create_post(post: PostCreate):
     post_to_add = PostOrm(
         title=post.title,
         description=post.description,
@@ -88,6 +98,11 @@ async def insert_post(post: PostCreate):
 async def update_post(post_id: int, post_data: PostUpdate):
     async with session_factory() as session:
         post = await session.get(PostOrm, post_id)
+        if post is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f"Post with ID: {post_id} not found"
+            )
         post.title = post_data.title or post.title
         post.description = post_data.description or post.description
         await session.commit()
@@ -98,6 +113,10 @@ async def update_post(post_id: int, post_data: PostUpdate):
 async def delete_user(user_id: int):
     async with session_factory() as session:
         user = await session.get(UserOrm, user_id)
+        if user is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f"User {user_id} not found")
         await session.delete(user)
         await session.commit()
         return user
@@ -106,15 +125,11 @@ async def delete_user(user_id: int):
 async def delete_post(post_id: int):
     async with session_factory() as session:
         post = await session.get(PostOrm, post_id)
+        if post is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f"Post with ID: {post_id} not found"
+            )
         await session.delete(post)
         await session.commit()
         return post
-
-
-async def authenticate_user(username: str, password: str):
-    user = await get_user_by_username(username)
-    if not user:
-        return None
-    if not verify_password(password, user["password"]):
-        return None
-    return user
